Add comprehensive validation and fix documentation tooling
- Add final_validation.sh with complete foundation testing
- Fix docs-setup CGO compilation issues
- Add comprehensive smoke_test.sh for CI validation
- Update documentation generation path references
- All systems validated and working correctly
