+++ docs/content/api/index.md
@@ -1,0 +1,280 @@
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# scripts
+
+```go
+import "github.com/your-org/go-template-project/scripts"
+```
+
+## Index
+
+- [type ProjectConfig](<#ProjectConfig>)
+- [type TemplateData](<#TemplateData>)
+
+
+<a name="ProjectConfig"></a>
+## type ProjectConfig
+
+ProjectConfig holds the configuration for project initialization.
+
+```go
+type ProjectConfig struct {
+    ProjectName    string
+    ModulePath     string
+    Description    string
+    Author         string
+    Email          string
+    License        string
+    EnableCLI      bool
+    EnableServer   bool
+    EnableWorker   bool
+    EnableDocs     bool
+    EnableE2ETests bool
+    GitRemote      string
+}
+```
+
+<a name="TemplateData"></a>
+## type TemplateData
+
+TemplateData holds data for template rendering.
+
+```go
+type TemplateData struct {
+    ProjectConfig
+    Year string
+}
+```
+
+# cli
+
+```go
+import "github.com/your-org/go-template-project/cmd/cli"
+```
+
+## Index
+
+
+
+# server
+
+```go
+import "github.com/your-org/go-template-project/cmd/server"
+```
+
+## Index
+
+
+
+# worker
+
+```go
+import "github.com/your-org/go-template-project/cmd/worker"
+```
+
+## Index
+
+- [type Worker](<#Worker>)
+  - [func NewWorker\(cfg \*config.Config\) \*Worker](<#NewWorker>)
+  - [func \(w \*Worker\) Start\(ctx context.Context\)](<#Worker.Start>)
+  - [func \(w \*Worker\) Stop\(\)](<#Worker.Stop>)
+
+
+<a name="Worker"></a>
+## type Worker
+
+Worker represents a background worker.
+
+```go
+type Worker struct {
+    // contains filtered or unexported fields
+}
+```
+
+<a name="NewWorker"></a>
+### func NewWorker
+
+```go
+func NewWorker(cfg *config.Config) *Worker
+```
+
+NewWorker creates a new worker instance.
+
+<a name="Worker.Start"></a>
+### func \(\*Worker\) Start
+
+```go
+func (w *Worker) Start(ctx context.Context)
+```
+
+Start begins the worker processing loop.
+
+<a name="Worker.Stop"></a>
+### func \(\*Worker\) Stop
+
+```go
+func (w *Worker) Stop()
+```
+
+Stop gracefully stops the worker.
+
+# app
+
+```go
+import "github.com/your-org/go-template-project/internal/app"
+```
+
+## Index
+
+- [type App](<#App>)
+  - [func New\(name, version string\) \*App](<#New>)
+  - [func \(a \*App\) GetInfo\(\) map\[string\]string](<#App.GetInfo>)
+  - [func \(a \*App\) Run\(\) error](<#App.Run>)
+
+
+<a name="App"></a>
+## type App
+
+App represents the core application.
+
+```go
+type App struct {
+    Name    string
+    Version string
+    Debug   bool
+}
+```
+
+<a name="New"></a>
+### func New
+
+```go
+func New(name, version string) *App
+```
+
+New creates a new application instance.
+
+<a name="App.GetInfo"></a>
+### func \(\*App\) GetInfo
+
+```go
+func (a *App) GetInfo() map[string]string
+```
+
+GetInfo returns basic application information.
+
+<a name="App.Run"></a>
+### func \(\*App\) Run
+
+```go
+func (a *App) Run() error
+```
+
+Run is the main entry point for CLI applications. Separated from main\(\) to make testing easier.
+
+# config
+
+```go
+import "github.com/your-org/go-template-project/internal/config"
+```
+
+## Index
+
+- [type Config](<#Config>)
+  - [func Load\(\) \(\*Config, error\)](<#Load>)
+  - [func \(c \*Config\) Address\(\) string](<#Config.Address>)
+
+
+<a name="Config"></a>
+## type Config
+
+Config holds application configuration.
+
+```go
+type Config struct {
+    Port         int           `json:"port"`
+    Host         string        `json:"host"`
+    Debug        bool          `json:"debug"`
+    ReadTimeout  time.Duration `json:"read_timeout"`
+    WriteTimeout time.Duration `json:"write_timeout"`
+    DatabaseURL  string        `json:"database_url,omitempty"`
+}
+```
+
+<a name="Load"></a>
+### func Load
+
+```go
+func Load() (*Config, error)
+```
+
+Load creates a new configuration from environment variables.
+
+<a name="Config.Address"></a>
+### func \(\*Config\) Address
+
+```go
+func (c *Config) Address() string
+```
+
+Address returns the full address to bind to.
+
+# handlers
+
+```go
+import "github.com/your-org/go-template-project/internal/handlers"
+```
+
+## Index
+
+- [func HealthCheck\(version string\) http.HandlerFunc](<#HealthCheck>)
+- [func ReadinessCheck\(\) http.HandlerFunc](<#ReadinessCheck>)
+- [type HealthResponse](<#HealthResponse>)
+
+
+<a name="HealthCheck"></a>
+## func HealthCheck
+
+```go
+func HealthCheck(version string) http.HandlerFunc
+```
+
+HealthCheck returns the application health status.
+
+GET /health
+
+Returns:
+
+- 200: Application is healthy
+- 503: Application has issues
+
+<a name="ReadinessCheck"></a>
+## func ReadinessCheck
+
+```go
+func ReadinessCheck() http.HandlerFunc
+```
+
+ReadinessCheck returns whether the application is ready to serve traffic.
+
+GET /ready
+
+Returns:
+
+- 200: Application is ready
+- 503: Application is not ready
+
+<a name="HealthResponse"></a>
+## type HealthResponse
+
+HealthResponse represents the health check response.
+
+```go
+type HealthResponse struct {
+    Status    string    `json:"status"`
+    Timestamp time.Time `json:"timestamp"`
+    Version   string    `json:"version,omitempty"`
+}
+```
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
