+++ docs/go.mod
@@ -1,5 +1,3 @@
 module github.com/your-org/go-template-project/docs
 
 go 1.22
-
-require github.com/google/docsy v0.10.0 // indirect
