+++ docs/layouts/_default/baseof.html
@@ -1,0 +1,90 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>{{ .Title }} | {{ .Site.Title }}</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <style>
+        body { 
+            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif; 
+            line-height: 1.6; 
+            max-width: 1200px; 
+            margin: 0 auto; 
+            padding: 20px;
+            color: #333;
+        }
+        nav ul { 
+            list-style: none; 
+            padding: 0; 
+            display: flex; 
+            background: #f5f5f5; 
+            margin: 0 0 20px 0;
+            border-radius: 5px;
+        }
+        nav li { 
+            margin-right: 20px; 
+        }
+        nav a { 
+            text-decoration: none; 
+            color: #007acc; 
+            display: block;
+            padding: 15px 20px;
+        }
+        nav a:hover { 
+            background: #e0e0e0; 
+            border-radius: 5px;
+        }
+        h1, h2, h3 { 
+            color: #2c3e50; 
+        }
+        code { 
+            background: #f8f8f8; 
+            padding: 2px 4px; 
+            border-radius: 3px; 
+            font-family: 'Monaco', 'Menlo', monospace;
+        }
+        pre { 
+            background: #f8f8f8; 
+            padding: 15px; 
+            border-radius: 5px; 
+            overflow-x: auto;
+        }
+        pre code { 
+            background: none; 
+            padding: 0; 
+        }
+        .content {
+            margin-top: 20px;
+        }
+        footer {
+            margin-top: 40px;
+            padding-top: 20px;
+            border-top: 1px solid #eee;
+            color: #666;
+            font-size: 0.9em;
+        }
+    </style>
+</head>
+<body>
+    <header>
+        <h1><a href="/" style="text-decoration: none; color: inherit;">{{ .Site.Title }}</a></h1>
+        {{ with .Site.Menus.main }}
+        <nav>
+            <ul>
+            {{ range . }}
+                <li><a href="{{ .URL }}">{{ .Name }}</a></li>
+            {{ end }}
+            </ul>
+        </nav>
+        {{ end }}
+    </header>
+    
+    <main class="content">
+        {{ block "main" . }}{{ end }}
+    </main>
+    
+    <footer>
+        <p>&copy; {{ now.Format "2006" }} {{ .Site.Title }}. Generated with <a href="https://gohugo.io">Hugo</a>.</p>
+    </footer>
+</body>
+</html>
