+++ docs/layouts/_default/list.html
@@ -1,0 +1,13 @@
+{{ define "main" }}
+<h1>{{ .Title }}</h1>
+{{ .Content }}
+
+{{ if .Pages }}
+<h2>Pages</h2>
+<ul>
+{{ range .Pages }}
+    <li><a href="{{ .RelPermalink }}">{{ .Title }}</a> - {{ .Summary }}</li>
+{{ end }}
+</ul>
+{{ end }}
+{{ end }}
