+++ docs/layouts/_default/single.html
@@ -1,0 +1,6 @@
+{{ define "main" }}
+<article>
+    <h1>{{ .Title }}</h1>
+    {{ .Content }}
+</article>
+{{ end }}
