+++ docs/public/docs/index.html
@@ -1,0 +1,15 @@
+<!doctype html><html><head><meta charset=utf-8><title>Documentation | Go Template Project</title><meta name=viewport content="width=device-width,initial-scale=1"><style>body{font-family:-apple-system,BlinkMacSystemFont,segoe ui,Roboto,sans-serif;line-height:1.6;max-width:1200px;margin:0 auto;padding:20px;color:#333}nav ul{list-style:none;padding:0;display:flex;background:#f5f5f5;margin:0 0 20px;border-radius:5px}nav li{margin-right:20px}nav a{text-decoration:none;color:#007acc;display:block;padding:15px 20px}nav a:hover{background:#e0e0e0;border-radius:5px}h1,h2,h3{color:#2c3e50}code{background:#f8f8f8;padding:2px 4px;border-radius:3px;font-family:monaco,menlo,monospace}pre{background:#f8f8f8;padding:15px;border-radius:5px;overflow-x:auto}pre code{background:0 0;padding:0}.content{margin-top:20px}footer{margin-top:40px;padding-top:20px;border-top:1px solid #eee;color:#666;font-size:.9em}</style></head><body><header><h1><a href=/ style=text-decoration:none;color:inherit>Go Template Project</a></h1><nav><ul><li><a href=/go-template-project/docs/>Documentation</a></li><li><a href=/go-template-project/api/>API Reference</a></li><li><a href=/go-template-project/examples/>Examples</a></li></ul></nav></header><main class=content><h1>Documentation</h1><h1 id=documentation>Documentation</h1><p>Welcome to the Go Template Project documentation. This section provides comprehensive guides for using and customizing the template.</p><h2 id=quick-navigation>Quick Navigation</h2><ul><li><strong><a href=getting-started/>Getting Started</a></strong> - Setup and initialization guide</li><li><strong><a href=architecture/>Architecture</a></strong> - Design patterns and project structure</li><li><strong><a href=development/>Development Workflow</a></strong> - Daily development practices</li><li><strong><a href=testing/>Testing Guide</a></strong> - Testing strategies and best practices</li><li><strong><a href=deployment/>Deployment</a></strong> - Docker and production deployment</li><li><strong><a href=configuration/>Configuration</a></strong> - Environment and application config</li></ul><h2 id=whats-included>What&rsquo;s Included</h2><p>The Go Template Project is a complete starting point for Go applications that includes:</p><h3 id=development-environment>Development Environment</h3><ul><li>Complete Go toolchain setup</li><li>Code formatting and linting</li><li>Pre-commit hooks for quality</li><li>Comprehensive Makefile</li></ul><h3 id=project-structure>Project Structure</h3><ul><li>Standard Go project layout</li><li>Modular architecture patterns</li><li>Example applications (CLI, server, worker)</li><li>Clear separation of concerns</li></ul><h3 id=quality-assurance>Quality Assurance</h3><ul><li>Progressive testing strategy</li><li>Security scanning and vulnerability checks</li><li>Code coverage reporting</li><li>Automated quality gates</li></ul><h3 id=deployment-ready>Deployment Ready</h3><ul><li>Multi-stage Docker builds</li><li>GitHub Actions CI/CD</li><li>Release automation</li><li>Production configurations</li></ul><h2 id=getting-help>Getting Help</h2><ul><li>Check the <a href=troubleshooting/>troubleshooting guide</a> for common issues</li><li>Review <a href=../examples/>examples</a> for code patterns</li><li>See <a href=../api/>API documentation</a> for package details</li><li>Open an issue on <a href=https://github.com/your-org/go-template-project/issues>GitHub</a></li></ul><h2>Pages</h2><ul><li><a href=/go-template-project/docs/getting-started/>Getting Started</a> -<h1 id=getting-started>Getting Started</h1><p>This guide will help you set up and customize the Go Template Project for your needs.</p><h2 id=prerequisites>Prerequisites</h2><p>Ensure you have the following installed:</p><ul><li><strong>Go 1.22+</strong> - <a href=https://golang.org/downloads/>Install from golang.org</a></li><li><strong>Git</strong> - For version control</li><li><strong>Make</strong> - For build automation</li><li><strong>Docker</strong> (optional) - For containerized development</li></ul><div class=highlight><div style=background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4><table style=border-spacing:0;padding:0;margin:0;border:0><tr><td style=vertical-align:top;padding:0;margin:0;border:0><pre tabindex=0 style=background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4><code><span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:.4em;padding:0 .4em;color:#7f7f7f">1
+</span><span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:.4em;padding:0 .4em;color:#7f7f7f">2
+</span><span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:.4em;padding:0 .4em;color:#7f7f7f">3
+</span><span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:.4em;padding:0 .4em;color:#7f7f7f">4
+</span><span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:.4em;padding:0 .4em;color:#7f7f7f">5
+</span></code></pre></td><td style=vertical-align:top;padding:0;margin:0;border:0;width:100%><pre tabindex=0 style=background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4><code class=language-bash data-lang=bash><span style=display:flex><span><span style=color:#57606a># Verify installations</span>
+</span></span><span style=display:flex><span>go version    <span style=color:#57606a># Should show Go 1.22+</span>
+</span></span><span style=display:flex><span>git --version
+</span></span><span style=display:flex><span>make --version
+</span></span><span style=display:flex><span>docker --version
+</span></span></code></pre></td></tr></table></div></div><h2 id=quick-setup>Quick Setup</h2><h3 id=1-clone-the-template>1. Clone the Template</h3><div class=highlight><div style=background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4><table style=border-spacing:0;padding:0;margin:0;border:0><tr><td style=vertical-align:top;padding:0;margin:0;border:0><pre tabindex=0 style=background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4><code><span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:.4em;padding:0 .4em;color:#7f7f7f">1
+</span><span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:.4em;padding:0 .4em;color:#7f7f7f">2
+</span></code></pre></td><td style=vertical-align:top;padding:0;margin:0;border:0;width:100%><pre tabindex=0 style=background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4><code class=language-bash data-lang=bash><span style=display:flex><span>git clone https://github.com/your-org/go-template-project.git my-awesome-project
+</span></span><span style=display:flex><span><span style=color:#6639ba>cd</span> my-awesome-project
+</span></span></code></pre></td></tr></table></div></div><h3 id=2-initialize-your-project>2. Initialize Your Project</h3><p>The interactive initialization script will customize the template:</p></li></ul></main><footer><p>&copy; 2025 Go Template Project. Generated with <a href=https://gohugo.io>Hugo</a>.</p></footer></body></html>
