+++ docs/public/docs/index.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="utf-8" standalone="yes"?><rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom"><channel><title>Documentation on Go Template Project</title><link>https://safurrier.github.io/go-template-project/docs/</link><description>Recent content in Documentation on Go Template Project</description><generator>Hugo</generator><language>en-us</language><atom:link href="https://safurrier.github.io/go-template-project/docs/index.xml" rel="self" type="application/rss+xml"/><item><title>Getting Started</title><link>https://safurrier.github.io/go-template-project/docs/getting-started/</link><pubDate>Mon, 01 Jan 0001 00:00:00 +0000</pubDate><guid>https://safurrier.github.io/go-template-project/docs/getting-started/</guid><description>&lt;h1 id="getting-started">Getting Started&lt;/h1>
+&lt;p>This guide will help you set up and customize the Go Template Project for your needs.&lt;/p>
+&lt;h2 id="prerequisites">Prerequisites&lt;/h2>
+&lt;p>Ensure you have the following installed:&lt;/p>
+&lt;ul>
+&lt;li>&lt;strong>Go 1.22+&lt;/strong> - &lt;a href="https://golang.org/downloads/">Install from golang.org&lt;/a>&lt;/li>
+&lt;li>&lt;strong>Git&lt;/strong> - For version control&lt;/li>
+&lt;li>&lt;strong>Make&lt;/strong> - For build automation&lt;/li>
+&lt;li>&lt;strong>Docker&lt;/strong> (optional) - For containerized development&lt;/li>
+&lt;/ul>
+&lt;div class="highlight">&lt;div style="background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4;">
+&lt;table style="border-spacing:0;padding:0;margin:0;border:0;">&lt;tr>&lt;td style="vertical-align:top;padding:0;margin:0;border:0;">
+&lt;pre tabindex="0" style="background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4;">&lt;code>&lt;span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:0.4em;padding:0 0.4em 0 0.4em;color:#7f7f7f">1
+&lt;/span>&lt;span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:0.4em;padding:0 0.4em 0 0.4em;color:#7f7f7f">2
+&lt;/span>&lt;span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:0.4em;padding:0 0.4em 0 0.4em;color:#7f7f7f">3
+&lt;/span>&lt;span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:0.4em;padding:0 0.4em 0 0.4em;color:#7f7f7f">4
+&lt;/span>&lt;span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:0.4em;padding:0 0.4em 0 0.4em;color:#7f7f7f">5
+&lt;/span>&lt;/code>&lt;/pre>&lt;/td>
+&lt;td style="vertical-align:top;padding:0;margin:0;border:0;;width:100%">
+&lt;pre tabindex="0" style="background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4;">&lt;code class="language-bash" data-lang="bash">&lt;span style="display:flex;">&lt;span>&lt;span style="color:#57606a"># Verify installations&lt;/span>
+&lt;/span>&lt;/span>&lt;span style="display:flex;">&lt;span>go version &lt;span style="color:#57606a"># Should show Go 1.22+&lt;/span>
+&lt;/span>&lt;/span>&lt;span style="display:flex;">&lt;span>git --version
+&lt;/span>&lt;/span>&lt;span style="display:flex;">&lt;span>make --version
+&lt;/span>&lt;/span>&lt;span style="display:flex;">&lt;span>docker --version
+&lt;/span>&lt;/span>&lt;/code>&lt;/pre>&lt;/td>&lt;/tr>&lt;/table>
+&lt;/div>
+&lt;/div>&lt;h2 id="quick-setup">Quick Setup&lt;/h2>
+&lt;h3 id="1-clone-the-template">1. Clone the Template&lt;/h3>
+&lt;div class="highlight">&lt;div style="background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4;">
+&lt;table style="border-spacing:0;padding:0;margin:0;border:0;">&lt;tr>&lt;td style="vertical-align:top;padding:0;margin:0;border:0;">
+&lt;pre tabindex="0" style="background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4;">&lt;code>&lt;span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:0.4em;padding:0 0.4em 0 0.4em;color:#7f7f7f">1
+&lt;/span>&lt;span style="white-space:pre;-webkit-user-select:none;user-select:none;margin-right:0.4em;padding:0 0.4em 0 0.4em;color:#7f7f7f">2
+&lt;/span>&lt;/code>&lt;/pre>&lt;/td>
+&lt;td style="vertical-align:top;padding:0;margin:0;border:0;;width:100%">
+&lt;pre tabindex="0" style="background-color:#fff;-moz-tab-size:4;-o-tab-size:4;tab-size:4;">&lt;code class="language-bash" data-lang="bash">&lt;span style="display:flex;">&lt;span>git clone https://github.com/your-org/go-template-project.git my-awesome-project
+&lt;/span>&lt;/span>&lt;span style="display:flex;">&lt;span>&lt;span style="color:#6639ba">cd&lt;/span> my-awesome-project
+&lt;/span>&lt;/span>&lt;/code>&lt;/pre>&lt;/td>&lt;/tr>&lt;/table>
+&lt;/div>
+&lt;/div>&lt;h3 id="2-initialize-your-project">2. Initialize Your Project&lt;/h3>
+&lt;p>The interactive initialization script will customize the template:&lt;/p></description></item></channel></rss>
