+++ docs/public/examples/index.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="utf-8" standalone="yes"?><rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom"><channel><title>Examples on Go Template Project</title><link>https://safurrier.github.io/go-template-project/examples/</link><description>Recent content in Examples on Go Template Project</description><generator>Hugo</generator><language>en-us</language><atom:link href="https://safurrier.github.io/go-template-project/examples/index.xml" rel="self" type="application/rss+xml"/></channel></rss>
