+++ docs/public/sitemap.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="utf-8" standalone="yes"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml"><url><loc>https://safurrier.github.io/go-template-project/docs/getting-started/</loc></url><url><loc>https://safurrier.github.io/go-template-project/docs/</loc></url><url><loc>https://safurrier.github.io/go-template-project/examples/</loc></url><url><loc>https://safurrier.github.io/go-template-project/api/</loc></url><url><loc>https://safurrier.github.io/go-template-project/categories/</loc></url><url><loc>https://safurrier.github.io/go-template-project/</loc></url><url><loc>https://safurrier.github.io/go-template-project/tags/</loc></url></urlset>
